requirements.txt
pandas
python-dotenv
requests
openpyxl
flask
wget
ibm_watson_machine_learning
ibm_watsonx_ai
ibm_cloud_sdk_core